Use thread variable for forcing master reads
The previous approach of unsetting and setting `default_slave_group` was not thread safe. We need an instance variable in order for it to be preserved across threads (I can’t find any way to tie into puma’s thread forking to set a `current_slave_group` after a fork). However we can’t be making changes to this variable across multiple threads and expecting it to be safe.

This approach uses a thread variable to say that a :master query has been requested, and then to clear it when the `current_slave_group` is reset back to what it was.